+++ src_for_visualstudio/DatabaseSystemProject/DatabaseSystemProject/MapReduce.cpp
@@ -1,11 +1,13 @@
-#include "MapReduce.h"
+#include "MapReduce.hpp"
 
-template <typename T> void ConcurrentQueue::push(T element) {
-	std::lock_guard<std::mutex> lock(mtx);
-	queue.push(std::move(element));
+template <typename T>
+void ConcurrentQueue<T>::push(T element) {
+    std::lock_guard<std::mutex> lock(mtx);
+    queue.push(std::move(element));
 }
 
-bool ConcurrentQueue::try_pop(T& popped_value) {
+template <typename T>
+bool ConcurrentQueue<T>::try_pop(T& popped_value) {
     std::lock_guard<std::mutex> lock(mtx);
     if (queue.empty()) {
         return false;
