--- src_for_visualstudio/DatabaseSystemProject/DatabaseSystemProject/input.txt
@@ -1,4 +1,0 @@
-How was your day? I was the best. How is the weather? It's so hot here.  What are you going to do? I will read a book.
-I'm not busy. Take your time, please. That's what I'm saying. I was really surprised! Is there a hotel in this area? I had it until last year.
-weather
-year
